fix(FreezeField): Fix header mismatch while rowspan in ajax pagination
Problem:
Table header is not aligned with table body when there are rowspan and colspan in ajax pagination function.
It is because the cell string 'Sub-Total' merged storage that makes the header not aligned with the body.

Solve:
It need to call AlignedHeader() in drawCallback() when ajax pagination function.
The AlignedHeader() function only redraw the table by PageName.
The left value of body and left value of header in FreezeField () should separate assign.